add some improvements and feedback
