+++ mqtt/mqtt-bridge/src/pump/builder.rs
@@ -16,14 +16,20 @@ use crate::{
 
 use super::{MessagesProcessor, Pump, PumpHandle};
 
-pub type PumpsResult = Result<
-    (
-        Pump<LocalUpstreamHandler<WakingMemoryStore>, LocalUpstreamPumpEventHandler>,
-        Pump<RemoteUpstreamHandler<WakingMemoryStore>, RemoteUpstreamPumpEventHandler>,
-    ),
-    BridgeError,
->;
+pub type PumpPair = (
+    Pump<LocalUpstreamHandler<WakingMemoryStore>, LocalUpstreamPumpEventHandler>,
+    Pump<RemoteUpstreamHandler<WakingMemoryStore>, RemoteUpstreamPumpEventHandler>,
+);
 
+/// Constructs a pair of bridge pumps: local and remote.
+///
+/// Local pump connects to a local broker, subscribes to topics to receive
+/// messages from local broker and put it in the store of the remote pump.
+/// Also reads messages from a local store and publishes them to local broker.
+///
+/// Remote pump connects to a remote broker, subscribes to topics to receive
+/// messages from remote broker and put it in the store of the local pump.
+/// Also reads messages from a remote store and publishes them to local broker.
 pub struct Builder {
     local: PumpBuilder,
     remote: PumpBuilder,
@@ -41,6 +47,7 @@ impl Default for Builder {
 }
 
 impl Builder {
+    /// Apples parameters to create local pump.
     pub fn with_local<F>(&mut self, mut apply: F) -> &mut Self
     where
         F: FnMut(&mut PumpBuilder),
@@ -49,6 +56,7 @@ impl Builder {
         self
     }
 
+    /// Applies parameters to create remote pump.
     pub fn with_remote<F>(&mut self, mut apply: F) -> &mut Self
     where
         F: FnMut(&mut PumpBuilder),
@@ -57,6 +65,7 @@ impl Builder {
         self
     }
 
+    /// Setups a factory to create publication store.
     pub fn with_store<F>(&mut self, store: F) -> &mut Self
     where
         F: Fn() -> PublicationStore<WakingMemoryStore> + 'static,
@@ -65,12 +74,10 @@ impl Builder {
         self
     }
 
-    pub fn build(&mut self) -> PumpsResult {
-        let ingress_store = (self.store)();
-        let ingress_loader = ingress_store.loader();
-
-        let egress_store = (self.store)();
-        let egress_loader = egress_store.loader();
+    /// Creates a pair of local and remote pump.
+    pub fn build(&mut self) -> Result<PumpPair, BridgeError> {
+        let remote_store = (self.store)();
+        let local_store = (self.store)();
 
         let (remote_messages_send, remote_messages_recv) = mpsc::channel(100);
         let (local_messages_send, local_messages_recv) = mpsc::channel(100);
@@ -78,7 +85,7 @@ impl Builder {
         let (subscriptions, topic_filters) = make_topics(&self.local.rules)?;
 
         let rpc = LocalRpcHandler::new(PumpHandle::new(remote_messages_send.clone()));
-        let messages = MessageHandler::new(ingress_store.clone(), topic_filters);
+        let messages = MessageHandler::new(remote_store.clone(), topic_filters);
         let handler = LocalUpstreamHandler::new(messages, rpc);
 
         let config = self.local.client.take().expect("local client config");
@@ -92,15 +99,14 @@ impl Builder {
             local_messages_send,
             client,
             subscriptions,
-            ingress_loader,
-            egress_store.clone(),
+            local_store.clone(),
             messages,
         )?;
 
         let (subscriptions, topic_filters) = make_topics(&self.remote.rules)?;
 
         let rpc = RemoteRpcHandler;
-        let messages = MessageHandler::new(egress_store, topic_filters);
+        let messages = MessageHandler::new(local_store, topic_filters);
         let handler = RemoteUpstreamHandler::new(messages, rpc);
 
         let config = self.local.client.take().expect("local client config");
@@ -114,8 +120,7 @@ impl Builder {
             remote_messages_send,
             client,
             subscriptions,
-            egress_loader,
-            ingress_store,
+            remote_store,
             messages,
         )?;
 
@@ -123,6 +128,7 @@ impl Builder {
     }
 }
 
+/// Collects parameters to construct `Pump`.
 #[derive(Default)]
 pub struct PumpBuilder {
     client: Option<MqttClientConfig>,
@@ -130,11 +136,13 @@ pub struct PumpBuilder {
 }
 
 impl PumpBuilder {
+    /// Applies default topic translation rules.
     pub fn with_rules(&mut self, rules: Vec<TopicRule>) -> &mut Self {
         self.rules = rules;
         self
     }
 
+    /// Applies MQTT client settings.
     pub fn with_config(&mut self, config: MqttClientConfig) -> &mut Self {
         self.client = Some(config);
         self
