+++ mqtt/mqtt-bridge/src/pump/egress.rs
@@ -1,0 +1,98 @@
+use futures_util::{
+    stream::{StreamExt, TryStreamExt},
+    FutureExt,
+};
+use tokio::{select, sync::oneshot};
+use tracing::{debug, error, info};
+
+use mqtt3::PublishHandle;
+
+use crate::persist::{PublicationStore, WakingMemoryStore};
+
+/// Handles the egress of received publications.
+///
+/// It loads messages from the local store and publishes them as MQTT messages
+/// to the broker. After acknowledgement is received from the broker it
+/// deletes publication from the store.
+pub(crate) struct Egress {
+    publish_handle: PublishHandle,
+    store: PublicationStore<WakingMemoryStore>,
+    shutdown_send: Option<oneshot::Sender<()>>,
+    shutdown_recv: oneshot::Receiver<()>,
+}
+
+impl Egress {
+    /// Creates a new instance of egress.
+    pub(crate) fn new(
+        publish_handle: PublishHandle,
+        store: PublicationStore<WakingMemoryStore>,
+    ) -> Egress {
+        let (shutdown_send, shutdown_recv) = oneshot::channel();
+
+        Self {
+            publish_handle,
+            store,
+            shutdown_send: Some(shutdown_send),
+            shutdown_recv,
+        }
+    }
+
+    /// Returns a shutdown handle of egress.
+    pub(crate) fn handle(&mut self) -> EgressShutdownHandle {
+        EgressShutdownHandle(self.shutdown_send.take())
+    }
+
+    /// Runs egress processing.
+    pub(crate) async fn run(self) {
+        let Egress {
+            mut publish_handle,
+            store,
+            shutdown_recv,
+            ..
+        } = self;
+
+        info!("starting egress publication processing...");
+
+        let mut shutdown = shutdown_recv.fuse();
+        let mut loader = store.loader().fuse();
+
+        loop {
+            select! {
+                _ = &mut shutdown => {
+                    info!(" received shutdown signal for egress messages");
+                    break;
+                }
+                maybe_publication = loader.try_next() => {
+                    debug!("extracted publication from store");
+
+                    if let Ok(Some((key, publication))) = maybe_publication {
+                        debug!("publishing {:?}", key);
+                        if let Err(e) = publish_handle.publish(publication).await {
+                            error!(err = %e, "failed publish");
+                        }
+
+                        if let Err(e) = store.remove(key) {
+                            error!(err = %e, "failed removing publication from store");
+                        }
+                    }
+                }
+            }
+        }
+
+        info!("finished egress publication processing");
+    }
+}
+
+/// Egress shutdown handle.
+pub(crate) struct EgressShutdownHandle(Option<oneshot::Sender<()>>);
+
+impl EgressShutdownHandle {
+    /// Sends a signal to shutdown egress.
+    pub(crate) async fn shutdown(mut self) {
+        if let Some(sender) = self.0.take() {
+            if sender.send(()).is_err() {
+                error!("unable to request shutdown for egress.");
+            }
+        }
+    }
+}
