+++ mqtt/mqtt-bridge/src/upstream/events/local.rs
@@ -5,12 +5,29 @@ use crate::{
     upstream::{CommandId, ConnectivityState},
 };
 
+/// Pump control event for a local upstream bridge pump.
 #[derive(Debug, PartialEq)]
 pub enum LocalUpstreamPumpEvent {
+    /// Connectivity update event.
     ConnectivityUpdate(ConnectivityState),
+
+    /// RPC command acknowledgement event.
     RpcAck(CommandId),
+
+    /// RPC command negative acknowledgement event.
+    RpcNack(CommandId),
 }
 
+/// Handles control event received by a local upstream bridge pump.
+///
+/// It handles follwing events:
+/// * connectivity update - emitted when the connection to remote broker changed
+///   (connected/disconnected). It should publish corresponding MQTT message to the
+///   local broker.
+/// * RPC command acknowledgement - emitted when the RPC command executed with
+///   success result.
+/// * RPC command negative acknowledgement - emitted when the RPC command failed
+///   to execute.
 pub struct LocalUpstreamPumpEventHandler;
 
 #[async_trait]
@@ -21,6 +38,7 @@ impl PumpMessageHandler for LocalUpstreamPumpEventHandler {
         match message {
             LocalUpstreamPumpEvent::ConnectivityUpdate(_) => {}
             LocalUpstreamPumpEvent::RpcAck(_) => {}
+            LocalUpstreamPumpEvent::RpcNack(_) => {}
         }
     }
 }
