+++ mqtt/mqtt-bridge/src/upstream/events/remote.rs
@@ -5,11 +5,17 @@ use crate::{
     upstream::{CommandId, RpcCommand},
 };
 
+/// Pump control event for a remote upstream bridge pump.
 #[derive(Debug, PartialEq)]
 pub enum RemoteUpstreamPumpEvent {
     RpcCommand(CommandId, RpcCommand),
 }
 
+/// Handles control event received by a remote upstream bridge pump.
+///
+/// It handles follwing events:
+/// * RPC command - emitted when `EdgeHub` requested RPC command to be executed
+/// against remote broker.
 pub struct RemoteUpstreamPumpEventHandler;
 
 #[async_trait]
