+++ mqtt/mqtt-bridge/src/upstream/rpc/mod.rs
@@ -1,5 +1,3 @@
-#![allow(dead_code)] // TODO remove when finished
-
 //! Downstream MQTT client event handler to react on RPC commands that
 //! `EdgeHub` sends to execute.
 //!
@@ -16,7 +14,7 @@ mod local;
 pub use local::LocalRpcHandler;
 pub use remote::RemoteRpcHandler;
 
-use std::sync::Arc;
+use std::{fmt::Display, fmt::Formatter, fmt::Result as FmtResult, sync::Arc};
 
 use bson::doc;
 use serde::{Deserialize, Serialize};
@@ -26,6 +24,7 @@ use mqtt3::PublishError;
 
 use crate::pump::PumpError;
 
+/// RPC command unique identificator.
 #[derive(Debug, Clone, PartialEq)]
 pub struct CommandId(Arc<String>);
 
@@ -38,12 +37,13 @@ where
     }
 }
 
-impl std::fmt::Display for CommandId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl Display for CommandId {
+    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
         write!(f, "{}", self.0)
     }
 }
 
+/// RPC command execution error.
 #[derive(Debug, thiserror::Error)]
 pub enum RpcError {
     #[error("failed to deserialize command from received publication")]
@@ -62,21 +62,25 @@ pub enum RpcError {
     SendToRemotePump(CommandId, #[source] PumpError),
 }
 
+/// RPC command to be executed against upstream broker.
 #[derive(Debug, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase", tag = "cmd")]
 pub enum RpcCommand {
+    /// A RPC command to subscribe to the topic.
     #[serde(rename = "sub")]
     Subscribe {
         #[serde(rename = "topic")]
         topic_filter: String,
     },
 
+    /// A RPC command to unsubscribe from the topic.
     #[serde(rename = "unsub")]
     Unsubscribe {
         #[serde(rename = "topic")]
         topic_filter: String,
     },
 
+    /// A RPC command to publish a message to a given topic.
     #[serde(rename = "pub")]
     Publish {
         topic: String,
